(25fc82c5f) Fix collider params being written as floats instead of ints.

+++ Barotrauma/BarotraumaClient/Source/Screens/CharacterEditorScreen.cs
@@ -4905,24 +4905,24 @@ namespace Barotrauma
                         // Capsules/Circles
                         //if (width == height)
                         //{
-                        //    colliderAttributes.Add(new XAttribute("radius", width / 2 * 0.85f));
+                        //    colliderAttributes.Add(new XAttribute("radius", (int)(width / 2 * 0.85f)));
                         //}
                         //else
                         //{
                         //    if (height > width)
                         //    {
-                        //        colliderAttributes.Add(new XAttribute("radius", width / 2 * 0.85f));
-                        //        colliderAttributes.Add(new XAttribute("height", height - width * 0.85f));
+                        //        colliderAttributes.Add(new XAttribute("radius", (int)(width / 2 * 0.85f)));
+                        //        colliderAttributes.Add(new XAttribute("height",(int) (height - width * 0.85f)));
                         //    }
                         //    else
                         //    {
-                        //        colliderAttributes.Add(new XAttribute("radius", height / 2 * 0.85f));
-                        //        colliderAttributes.Add(new XAttribute("width", width - height * 0.85f));
+                        //        colliderAttributes.Add(new XAttribute("radius", (int)(height / 2 * 0.85f)));
+                        //        colliderAttributes.Add(new XAttribute("width", (int)(width - height * 0.85f)));
                         //    }
                         //}
                         // Rectangles
-                        colliderAttributes.Add(new XAttribute("height", height * 0.85f));
-                        colliderAttributes.Add(new XAttribute("width", width * 0.85f));
+                        colliderAttributes.Add(new XAttribute("height", (int)(height * 0.85f)));
+                        colliderAttributes.Add(new XAttribute("width", (int)(width * 0.85f)));
                         idToCodeName.TryGetValue(id, out string notes);
                         LimbXElements.Add(id.ToString(), new XElement("limb",
                             new XAttribute("id", id),
